XWIKI-15912: Improper "rename log" string is displayed when user is performing a page copy
  * Create a dedicated CopyJob
