+++ xwiki-platform-core/xwiki-platform-refactoring/xwiki-platform-refactoring-api/src/main/java/org/xwiki/refactoring/internal/job/AbstractEntityJobWithChecks.java
@@ -51,12 +51,12 @@ public abstract class AbstractEntityJobWithChecks<R extends EntityRequest, S ext
     protected Map<EntityReference, EntitySelection> concernedEntities = new HashMap<>();
 
     /**
-     * @return true means that the deletion check will be skipped.
+     * @return true means that this job will delete sources.
      * @since 10.11RC1
      */
-    protected boolean isSkipChecks()
+    protected boolean isDeleteSources()
     {
-        return false;
+        return true;
     }
 
     @Override
@@ -70,7 +70,7 @@ public abstract class AbstractEntityJobWithChecks<R extends EntityRequest, S ext
                 progressManager.startStep(this);
                 getEntities(entityReferences);
 
-                if (!isSkipChecks()) {
+                if (isDeleteSources()) {
                     // Send the event
                     DocumentsDeletingEvent event = new DocumentsDeletingEvent();
                     observationManager.notify(event, this, concernedEntities);
