+++ xwiki-platform-core/xwiki-platform-refactoring/xwiki-platform-refactoring-api/src/main/java/org/xwiki/refactoring/internal/job/RenameJob.java
@@ -46,6 +46,13 @@ public class RenameJob extends MoveJob
     }
 
     @Override
+    protected boolean isDeleteSources()
+    {
+        // rename always delete original sources
+        return true;
+    }
+
+    @Override
     protected void process(Collection<EntityReference> entityReferences)
     {
         if (entityReferences.size() == 1) {
