+++ xwiki-platform-core/xwiki-platform-refactoring/xwiki-platform-refactoring-api/src/main/resources/ApplicationResources.properties
@@ -52,9 +52,10 @@ job.question.button.refactoring/rename.confirm=Rename
 job.question.button.refactoring/move.confirm=Move
 
 job.log.label.refactoring/rename=Rename log
-job.log.label.refactoring/rename=Move log
+job.log.label.refactoring/move=Move log
 job.log.label.refactoring/delete=Delete log
 job.log.label.refactoring/copyAs=Copy log
+job.log.label.refactoring/copy=Copy log
 
 job.question.OverwriteQuestion.title=Replace page "{0}"?
 job.question.OverwriteQuestion.message=An older page with the same name already exists in "{0}". Replacing it will overwrite its content.
